SAGE OPTIONAL :D

+++ gettools.sh
@@ -24,6 +24,10 @@ while [[ "${#}" -gt 0 && ."${1}" == .-* ]]; do
       MIRROR="${1}"; shift;;
     -m=*|--mirror=* )
       MIRROR="${opt#*=}";;
+    -s|--sage )
+      SAGE="${1}"; shift;;
+    -s=*|--sage=* )
+      SAGE="${opt#*=}";;
     *) echo -e ' '${RED}'[!]'${RESET}" Unknown option: ${RED}${x}${RESET}" 1>&2 \
       && exit 1;;
    esac
@@ -564,15 +568,17 @@ fi
 
 # sage
 # we don't want to re-install if it's already installed, 1.2gb download is kinda brutal
-if [ -e /usr/local/bin/sage ]; then
-  echo "$(date '+%X') ${YELLOW}[+]${RESET} Sage installed already"
-else
-  echo -e "$(date '+%X') ${GREEN}[+]${RESET} Installing sage... this may take a while!"
-  wget -q --show-progress http://mirror.aarnet.edu.au/pub/sage/linux/64bit/sage-7.6-Debian_GNU_Linux_8-x86_64.tar.bz2
-  tar jxf sage-7.6-Debian_GNU_Linux_8-x86_64.tar.bz2
-  mv SageMath/ /opt/
-  rm sage*
-  ln -s /opt/SageMath/sage /usr/local/bin/sage
+if [ ${SAGE} ]; then
+  if [ -e /usr/local/bin/sage ]; then
+    echo "$(date '+%X') ${YELLOW}[+]${RESET} Sage installed already"
+  else
+    echo -e "$(date '+%X') ${GREEN}[+]${RESET} Installing sage... this may take a while!"
+    wget -q --show-progress http://mirror.aarnet.edu.au/pub/sage/linux/64bit/sage-7.6-Debian_GNU_Linux_8-x86_64.tar.bz2
+    tar jxf sage-7.6-Debian_GNU_Linux_8-x86_64.tar.bz2
+    mv SageMath/ /opt/
+    rm sage*
+    ln -s /opt/SageMath/sage /usr/local/bin/sage
+  fi
 fi
 
 # dnsmasq
